Detect separate username and password fields
The current logic assumes that username and password fields will always be
rendered in the DOM at the same time, even if one is hidden (e.g. in a
multi-step login form). However, some forms only render them separately
(removing username from DOM when password field is appended), or they
change the username field to a <div> or <span>, which then escapes our
logic.

This commit attempts to capture username and password fields separately
so that we can support this type of form behaviour.